Fix Proj07 state machine bugs
- main.c: NUDGE_HOME branch built a local 'home' packet but never
  applied it; gimbal never received the home command. Write directly
  into ctl so hamfly_send_control actually transmits.
- main.c: route NUDGE_HOME mode through state_nudge_home (was
  unreachable). Replace PI_SERIAL_TEST -> state_error call with an
  explicit set_error so the error message is meaningful.
- app_ctx.h: home_entry_ms widened uint8_t -> uint32_t to match
  g_tick_ms assignment and 5000 ms timeout comparison.

+++ test_projects/Proj07_hamflyAPI_stateMachineUpgrade.cydsn/app_ctx.h
@@ -49,7 +49,7 @@ typedef struct {
     float    nudge_tilt_rate;
     uint8_t  nudge_time_short;
     uint8_t  nudge_time_long;
-    uint8_t  home_entry_ms;  // Timer for homing
+    uint32_t home_entry_ms;  // Timer for homing
  
     float    origin_pan_deg;
     float    origin_tilt_deg;
